+++ Source/C/GnuGenBootSector/GnuGenBootSector.c
@@ -396,7 +396,7 @@ main (
         return 1;
       }
       if (LogLevel > 9) {
-        Error (NULL, 0, 1003, "Invalid option value", "Debug Level range is 0-9, currnt input level is %d", LogLevel);
+        Error (NULL, 0, 1003, "Invalid option value", "Debug Level range is 0-9, currnt input level is %d", (int) LogLevel);
         return 1;
       }
       SetPrintLevel (LogLevel);
